+++ rbac-distributed-service-impl/src/main/java/com/dwarfeng/rbacds/impl/dao/preset/UserPresetCriteriaMaker.java
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Component;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Component
@@ -20,6 +21,9 @@ public class UserPresetCriteriaMaker implements PresetCriteriaMaker {
         switch (s) {
             case UserMaintainService.ID_LIKE:
                 idLike(detachedCriteria, objects);
+                break;
+            case UserMaintainService.CHILD_FOR_ROLE:
+                childForRole(detachedCriteria, objects);
                 break;
             case UserMaintainService.CHILD_FOR_ROLE_SET:
                 childForRoleSet(detachedCriteria, objects);
@@ -33,6 +37,20 @@ public class UserPresetCriteriaMaker implements PresetCriteriaMaker {
         try {
             String id = (String) objects[0];
             detachedCriteria.add(Restrictions.like("stringId", id, MatchMode.ANYWHERE));
+        } catch (Exception e) {
+            throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
+        }
+    }
+
+    private void childForRole(DetachedCriteria detachedCriteria, Object[] objects) {
+        try {
+            if (Objects.isNull(objects[0])) {
+                detachedCriteria.add(Restrictions.isEmpty("roles"));
+            } else {
+                StringIdKey stringIdKey = (StringIdKey) objects[0];
+                detachedCriteria.createAlias("roles", "r");
+                detachedCriteria.add(Restrictions.eq("r.stringId", stringIdKey.getStringId()));
+            }
         } catch (Exception e) {
             throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
         }