+++ rbac-distributed-service-impl/src/test/java/com/dwarfeng/rbacds/impl/service/UserMaintainServiceImplTest.java
@@ -6,6 +6,7 @@ import com.dwarfeng.rbacds.stack.service.RoleMaintainService;
 import com.dwarfeng.rbacds.stack.service.UserMaintainService;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
+import org.apache.commons.beanutils.BeanUtils;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -15,7 +16,12 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.Arrays;
+import java.util.List;
 import java.util.Objects;
+import java.util.stream.Collectors;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(locations = "classpath:spring/application-context*.xml")
@@ -29,6 +35,7 @@ public class UserMaintainServiceImplTest {
     private User zhangSan;
     private User liSi;
     private User wangWu;
+    private User zhaoLiu;
 
     private Role admin;
     private Role guest;
@@ -39,6 +46,7 @@ public class UserMaintainServiceImplTest {
         zhangSan = new User(new StringIdKey("zhang_san"), "测试用账号");
         liSi = new User(new StringIdKey("li_si"), "测试用账号");
         wangWu = new User(new StringIdKey("wang_wu"), "测试用账号");
+        zhaoLiu = new User(new StringIdKey("zhao_liu"), "测试用账号");
         admin = new Role(new StringIdKey("admin"), "管理员", true, "测试用角色");
         moderator = new Role(new StringIdKey("moderator"), "操作员", true, "测试用角色");
         guest = new Role(new StringIdKey("guest"), "访客", false, "测试用角色");
@@ -49,23 +57,51 @@ public class UserMaintainServiceImplTest {
         zhangSan = null;
         liSi = null;
         wangWu = null;
+        zhaoLiu = null;
         admin = null;
         guest = null;
         moderator = null;
     }
 
     @Test
-    public void test() throws ServiceException {
+    public void testForCrud() throws Exception {
+        try {
+            userMaintainService.insertOrUpdate(zhangSan);
+
+            User testZhangSan = userMaintainService.get(zhangSan.getKey());
+            assertEquals(BeanUtils.describe(zhangSan), BeanUtils.describe(testZhangSan));
+            userMaintainService.update(zhangSan);
+            testZhangSan = userMaintainService.get(zhangSan.getKey());
+            assertEquals(BeanUtils.describe(zhangSan), BeanUtils.describe(testZhangSan));
+        } finally {
+            userMaintainService.deleteIfExists(zhangSan.getKey());
+        }
+    }
+
+    @Test
+    public void testForLookup() throws ServiceException {
         try {
             userMaintainService.insertOrUpdate(zhangSan);
             userMaintainService.insertOrUpdate(liSi);
             userMaintainService.insertOrUpdate(wangWu);
+            userMaintainService.insertOrUpdate(zhaoLiu);
             roleMaintainService.insertOrUpdate(admin);
             roleMaintainService.insertOrUpdate(guest);
             roleMaintainService.insertOrUpdate(moderator);
             userMaintainService.batchAddRoleRelations(zhangSan.getKey(), Arrays.asList(admin.getKey(), moderator.getKey(), guest.getKey()));
             userMaintainService.batchAddRoleRelations(liSi.getKey(), Arrays.asList(moderator.getKey(), guest.getKey()));
             userMaintainService.addRoleRelation(wangWu.getKey(), guest.getKey());
+
+            List<StringIdKey> lookupUserKeys = userMaintainService.lookup(UserMaintainService.CHILD_FOR_ROLE, new Object[]{null})
+                    .getData().stream().map(User::getKey).collect(Collectors.toList());
+            assertTrue(lookupUserKeys.size() >= 1);
+            assertTrue(lookupUserKeys.contains(zhaoLiu.getKey()));
+            lookupUserKeys = userMaintainService.lookup(UserMaintainService.CHILD_FOR_ROLE, new Object[]{guest.getKey()})
+                    .getData().stream().map(User::getKey).collect(Collectors.toList());
+            assertTrue(lookupUserKeys.size() >= 3);
+            assertTrue(lookupUserKeys.contains(zhangSan.getKey()));
+            assertTrue(lookupUserKeys.contains(liSi.getKey()));
+            assertTrue(lookupUserKeys.contains(wangWu.getKey()));
         } finally {
             if (Objects.nonNull(zhangSan)) userMaintainService.deleteIfExists(zhangSan.getKey());
             roleMaintainService.deleteIfExists(admin.getKey());